backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
httpx==0.27.2
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
python-Levenshtein==0.26.1
rapidfuzz==3.10.1
nltk==3.9.1
transliterate==1.10.2
sentence-transformers==3.3.1
faiss-cpu==1.9.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
asyncpg==0.30.0
unidecode==1.3.8
python-multipart==0.0.18
aiofiles==24.1.0